feat: :sparkles: issuesコマンドの機能拡張とコード構造改善 #8
- issues編集機能を追加（edit サブコマンド）
- issueのURL表示機能を追加（url サブコマンド）
- 大きなissues.goファイルを機能別に分割
- クライアントにissue更新機能とデータ構造を追加
- 構造体フィールドのインデント統一
